리뷰 상세보기 데이터 반영 마무리

+++ app/src/main/java/com/nepplus/apipractice_20220106/models/ReviewData.kt
@@ -6,7 +6,7 @@ class ReviewData(
     val id : Int,
     val title : String,
     val content : String,
-    val score : Double,  // 파이썬에서는 float개념과 동일
+    val score : Float,  // 파이썬에서는 float개념과 동일
     val thumbnail_img : String,
     val user : UserData  // 기존에 만든 UserData 재활용
 ) :Serializable {
